+++ src/pdf.js
@@ -12,7 +12,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/* eslint-disable sort-exports/sort-exports */
 
 // eslint-disable-next-line max-len
 /** @typedef {import("./display/api").PDFDocumentLoadingTask} PDFDocumentLoadingTask */
@@ -20,19 +19,6 @@
 /** @typedef {import("./display/api").PDFPageProxy} PDFPageProxy */
 /** @typedef {import("./display/api").RenderTask} RenderTask */
 
-import {
-  addLinkAttributes,
-  getFilenameFromUrl,
-  getPdfFilenameFromUrl,
-  getXfaPageViewport,
-  isPdfFile,
-  isValidFetchUrl,
-  LinkTarget,
-  loadScript,
-  PDFDateString,
-  PixelsPerInch,
-  RenderingCancelledException,
-} from "./display/display_utils.js";
 import {
   AnnotationMode,
   CMapCompressionType,
@@ -60,6 +46,17 @@ import {
   setPDFNetworkStreamFactory,
   version,
 } from "./display/api.js";
+import {
+  getFilenameFromUrl,
+  getPdfFilenameFromUrl,
+  getXfaPageViewport,
+  isPdfFile,
+  isValidFetchUrl,
+  loadScript,
+  PDFDateString,
+  PixelsPerInch,
+  RenderingCancelledException,
+} from "./display/display_utils.js";
 import { AnnotationLayer } from "./display/annotation_layer.js";
 import { GlobalWorkerOptions } from "./display/worker_options.js";
 import { isNodeJS } from "./shared/is_node.js";
@@ -108,49 +105,39 @@ if (typeof PDFJSDev === "undefined" || !PDFJSDev.test("PRODUCTION")) {
 }
 
 export {
-  // From "./display/display_utils.js":
-  addLinkAttributes,
-  getFilenameFromUrl,
-  getPdfFilenameFromUrl,
-  isPdfFile,
-  LinkTarget,
-  loadScript,
-  PDFDateString,
-  PixelsPerInch,
-  RenderingCancelledException,
-  getXfaPageViewport,
-  // From "./shared/util.js":
+  AnnotationLayer,
   AnnotationMode,
+  build,
   CMapCompressionType,
   createObjectURL,
   createPromiseCapability,
   createValidAbsoluteUrl,
+  getDocument,
+  getFilenameFromUrl,
+  getPdfFilenameFromUrl,
+  getXfaPageViewport,
+  GlobalWorkerOptions,
   InvalidPDFException,
+  isPdfFile,
+  loadScript,
+  LoopbackPort,
   MissingPDFException,
   OPS,
   PasswordResponses,
+  PDFDataRangeTransport,
+  PDFDateString,
+  PDFWorker,
   PermissionFlag,
+  PixelsPerInch,
   removeNullCharacters,
+  RenderingCancelledException,
+  renderTextLayer,
   shadow,
+  SVGGraphics,
   UnexpectedResponseException,
   UNSUPPORTED_FEATURES,
   Util,
   VerbosityLevel,
-  // From "./display/api.js":
-  build,
-  getDocument,
-  LoopbackPort,
-  PDFDataRangeTransport,
-  PDFWorker,
   version,
-  // From "./display/annotation_layer.js":
-  AnnotationLayer,
-  // From "./display/worker_options.js":
-  GlobalWorkerOptions,
-  // From "./display/text_layer.js":
-  renderTextLayer,
-  // From "./display/svg.js":
-  SVGGraphics,
-  // From "./display/xfa_layer.js":
   XfaLayer,
 };
